Make your own review appear on spell page Change star Intfield to Floatfield Add star_list to spell detail page context

+++ spezspellz/models.py
@@ -143,8 +143,8 @@ class Review(models.Model):
     user: models.ForeignKey[User] = models.ForeignKey(User, null=False, on_delete=models.CASCADE)
     spell: models.ForeignKey[Spell] = models.ForeignKey(Spell, null=False, on_delete=models.CASCADE)
     review_desc: models.CharField = models.CharField(max_length=256, null=True)
-    star: models.IntegerField = models.IntegerField()
-    # lowest 1 star highest 5 star
+    star: models.FloatField = models.FloatField()
+    # lowest 0.5 star highest 5 star
 
     def __str__(self):
         """Return a review for a spell"""
